Add macro spectrotomo
A new data collection macro used to generate a txt TXM script,
is needed in Mistral. The name of the new macro is spectrotomo.

It will be used to collect images at different angles, and at
each angle we will vary the energy. This is useful in
spectrotomography, when it is necessary to have a really precise
alignment between all the images at different energies corresponding
to the same angle. By scaning through all the energies per each single
angle, we improve the alignment for each angle, that otherwise would
be lost if for each energy we were scaning through all the angles.