Modify file names to be able to preprocess
Modify file names and adapt them to be able to apply the
preprocessing program to convert from xrm to hdf5: xrm2nexus.

Apply the following patterns for the spectroscopy
images (energy scans):

- For Sample:
date_spectrumname_0_energy_0.xrm

- For FF:
date_spectrumname_0_FF_energy.xrm

In this way, we achieve to use the same programs used till now
in converting from xrm to hdf5 the tomography files, which where
using different ZP postions, different angles, and different
positions.

Now we are interested into storing all the images for different
energies in a single hdf5 file. Done. In this kind of
spectrocopies, a single angle is used.

Thus, the angle and the energy have been switched in the
pattern name, to be able to allow applying xrm2nexus with
spectrocopy experiments: images taken at a single angle,
multiple energies, and stored in a single hdf5 file.