Use DATE and NAME to configure the sample name
Use DATE and NAME to configure the sample name as:
DATE_NAME_

Make a verification than DATE and NAME does not contain
underscores. This is important for the posterior preprocessing.
This avoids failures by uncorrectly writting the name by
the users of the macro and the preprocessing, minimizing
failures at the preprocessing moment.