New acquisition macro for auto tomography
Develop new acquisition macro autotomo, based on manytomos, but
adding triggers (movement of motors prx and pry), that will
trigger a Device Server which will trigger the preprocessing
of the raw data after the acquisition of the whole projections
of each sample.

This macro creates the txt that is loaded in the TXM
microscope for projection acquisition. This txt shall also
be placed in the raw data folder, because is also used
for creating the DB where each record represents a xrm single
image: the DB is required for the preprocessing.