Set the txm_file and start the DS
Set the txm_file and start the DS TXMAutoProcessing, in order to
change its state to ON.

Thanks to that, the user can already load the txt file into
the microscope.

The macro autotomo cannot be launched till the DS is on Standby
again; otherwise, this would overwrite the TXM_file, and the
commands would be incorrect.