Adapt autotomo to distribute images in folders
Adapt autotomo macro in order to allow distribution of acquired
images by TXM Microscope in different folders.

This is required because in some cases the number of acquired
xrm image files per folder is really high (it can go from few image
files to 10000 image files in some cases). It becomes then
important to be able to distribute the high number of images in
different folders; at least to avoid slow readings of the files;
but also to avoid problems in the file system.