Allow a single sample for magnetism workflow
Allow only a single sample for magnetism workflow.
Magnetism workflow, currently, allows only a single sample,
because, a manual positioning of the sample must be done
manually by the scientists after the FF acquisition. Then,
FF txt must be loaded in TXM, then a manual positioning must be done,
and then target sample collect images txt must be loaded in TXM.

Because of this reason, multiple samples cannot be described in the
same txt file.

Modify macro to allow only a single sample in macro.